+++ main.go
@@ -19,6 +19,8 @@ func main() {
 	maxPlayers := flag.Int("max-players", 100, "Maximum Players in the server.")
 	origin := flag.String("origin", "https://demo.rabbitsky.io", "Hostname of the website you will serve the Static HTML. Please do remember to insert http:// or https:// for this to works.")
 	serverPassword := flag.String("admin-password", "", "Admin password to allow user to use admin command. Set on chat using '/admin [password]'. Leaving it empty will make command unusable.")
+	limitPosMin := flag.String("limit-position-min", "0,0,0", "Position (X,Y,Z) Minimum that user allowed. If user position is less than this value, they will be disconnected.")
+	limitPosMax := flag.String("limit-position-max", "4000,100,3000", "Position (X,Y,Z) Maximum that user allowed. If user position is more than this value, they will be disconnected.")
 	addBots := flag.Int("add-bots", 0, "Spawn this number of bots to debug.")
 	flag.Parse()
 
@@ -38,7 +40,7 @@ func main() {
 	go rsBot.AddBot(channel, *addBots, *serverTick)
 
 	/* Init HTTP Handler */
-	httpHandler, err := rsHTTPHandler.Init(channel, webSocket, *origin, *serverPassword)
+	httpHandler, err := rsHTTPHandler.Init(channel, webSocket, *origin, *serverPassword, *limitPosMin, *limitPosMax)
 	if err != nil {
 		log.Fatalln("[Error] Could not initialize WebSocket. Err: ", err)
 	}